Refactor some files to move out SideBar and prepare future work
