+++ package.json
@@ -7,6 +7,7 @@
     "antd": "^4.7.0",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
+    "react-firebaseui": "^4.1.0",
     "react-scripts": "3.4.3"
   },
   "scripts": {
